chore: move triewriter trait
